refactor: restructure test directory with conftest hierarchy
Restructure flat test directory to mirror source layout (models/,
repositories/, services/) with proper conftest.py hierarchy. Centralize
7 duplicated factory functions into 4 unified make_*() functions in
root conftest.py. Move mock repositories from helpers.py to conftest
files. Extract write_yaml() to repositories/conftest.py and pre-wired
catalog fixtures to services/conftest.py.

350 tests pass, 97.54% coverage, ruff clean, mypy improved (62→58).

Relates to #30
