+++ tests/models/conftest.py
@@ -1,0 +1,7 @@
+"""Model-layer test configuration.
+
+Models use plain factory functions from the root conftest — no stateful
+fixtures needed at this layer.
+"""
+
+from __future__ import annotations
