+++ tests/models/test_extract_config_type.py
