+++ tests/models/test_template_entry.py
