+++ tests/models/test_tool_entry.py
