+++ tests/repositories/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for catalog repository layer."""
