+++ tests/repositories/test_yaml_template_repo.py
@@ -3,28 +3,19 @@
 from pathlib import Path
 
 import pytest
-import yaml
 
 from akgentic.catalog.models.errors import CatalogValidationError, EntryNotFoundError
 from akgentic.catalog.models.queries import TemplateQuery
 from akgentic.catalog.models.template import TemplateEntry
 from akgentic.catalog.repositories.yaml.template_repo import YamlTemplateCatalogRepository
-
-
-def _write_yaml(path: Path, data: object) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(
-        yaml.dump(data, default_flow_style=False, sort_keys=False, allow_unicode=True),
-        encoding="utf-8",
-    )
-
+from tests.repositories.conftest import write_yaml
 
 # ---- Loading ----
 
 
 def test_load_single_yaml_file(tmp_path: Path) -> None:
     """AC #1: Load entries from a single YAML file containing list of dicts."""
-    _write_yaml(
+    write_yaml(
         tmp_path / "templates.yaml",
         [
             {"id": "t1", "template": "Hello {name}"},
@@ -39,8 +30,8 @@ def test_load_single_yaml_file(tmp_path: Path) -> None:
 
 def test_load_multiple_yaml_files(tmp_path: Path) -> None:
     """AC #1: Multiple YAML files merged into single registry."""
-    _write_yaml(tmp_path / "coordinator.yaml", [{"id": "c1", "template": "Coord {role}"}])
-    _write_yaml(tmp_path / "specialist.yaml", [{"id": "s1", "template": "Spec {skill}"}])
+    write_yaml(tmp_path / "coordinator.yaml", [{"id": "c1", "template": "Coord {role}"}])
+    write_yaml(tmp_path / "specialist.yaml", [{"id": "s1", "template": "Spec {skill}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     entries = repo.list()
     assert len(entries) == 2
@@ -50,8 +41,8 @@ def test_load_multiple_yaml_files(tmp_path: Path) -> None:
 
 def test_duplicate_id_across_files_raises(tmp_path: Path) -> None:
     """AC #2: Duplicate id across files raises CatalogValidationError with file paths."""
-    _write_yaml(tmp_path / "file_a.yaml", [{"id": "dup", "template": "A {x}"}])
-    _write_yaml(tmp_path / "file_b.yaml", [{"id": "dup", "template": "B {y}"}])
+    write_yaml(tmp_path / "file_a.yaml", [{"id": "dup", "template": "A {x}"}])
+    write_yaml(tmp_path / "file_b.yaml", [{"id": "dup", "template": "B {y}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     with pytest.raises(CatalogValidationError) as exc_info:
         repo.list()
@@ -81,7 +72,7 @@ def test_empty_yaml_file(tmp_path: Path) -> None:
 
 def test_single_dict_yaml_normalized_to_list(tmp_path: Path) -> None:
     """Edge case: YAML file with single dict (not list) is normalized."""
-    _write_yaml(tmp_path / "single.yaml", {"id": "solo", "template": "Solo {x}"})
+    write_yaml(tmp_path / "single.yaml", {"id": "solo", "template": "Solo {x}"})
     repo = YamlTemplateCatalogRepository(tmp_path)
     entries = repo.list()
     assert len(entries) == 1
@@ -93,7 +84,7 @@ def test_single_dict_yaml_normalized_to_list(tmp_path: Path) -> None:
 
 def test_get_returns_entry_by_id(tmp_path: Path) -> None:
     """AC #4: get() returns entry by id."""
-    _write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
+    write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     entry = repo.get("t1")
     assert entry is not None
@@ -102,7 +93,7 @@ def test_get_returns_entry_by_id(tmp_path: Path) -> None:
 
 def test_get_returns_none_for_missing(tmp_path: Path) -> None:
     """get() returns None for non-existent id."""
-    _write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
+    write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     assert repo.get("missing") is None
 
@@ -112,7 +103,7 @@ def test_get_returns_none_for_missing(tmp_path: Path) -> None:
 
 def test_list_returns_all_entries(tmp_path: Path) -> None:
     """list() returns all cached entries."""
-    _write_yaml(
+    write_yaml(
         tmp_path / "t.yaml",
         [
             {"id": "a", "template": "A {x}"},
@@ -129,7 +120,7 @@ def test_list_returns_all_entries(tmp_path: Path) -> None:
 
 def test_search_by_placeholder(tmp_path: Path) -> None:
     """AC #5: search(TemplateQuery(placeholder='role')) returns only matching templates."""
-    _write_yaml(
+    write_yaml(
         tmp_path / "t.yaml",
         [
             {"id": "t1", "template": "Hello {role} and {name}"},
@@ -146,7 +137,7 @@ def test_search_by_placeholder(tmp_path: Path) -> None:
 
 def test_search_by_id(tmp_path: Path) -> None:
     """search(TemplateQuery(id='foo')) exact id match."""
-    _write_yaml(
+    write_yaml(
         tmp_path / "t.yaml",
         [
             {"id": "foo", "template": "Foo {x}"},
@@ -161,7 +152,7 @@ def test_search_by_id(tmp_path: Path) -> None:
 
 def test_search_and_semantics(tmp_path: Path) -> None:
     """search(TemplateQuery(id='foo', placeholder='role')) AND semantics."""
-    _write_yaml(
+    write_yaml(
         tmp_path / "t.yaml",
         [
             {"id": "foo", "template": "Foo {role}"},
@@ -177,7 +168,7 @@ def test_search_and_semantics(tmp_path: Path) -> None:
 
 def test_search_no_filters_returns_all(tmp_path: Path) -> None:
     """search with all None fields returns all entries."""
-    _write_yaml(
+    write_yaml(
         tmp_path / "t.yaml",
         [
             {"id": "a", "template": "A {x}"},
@@ -194,13 +185,13 @@ def test_search_no_filters_returns_all(tmp_path: Path) -> None:
 
 def test_caching_no_reread(tmp_path: Path) -> None:
     """AC #4: second list() doesn't re-read files."""
-    _write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
+    write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     entries1 = repo.list()
     assert len(entries1) == 1
 
     # Modify file on disk
-    _write_yaml(
+    write_yaml(
         tmp_path / "t.yaml",
         [
             {"id": "t1", "template": "Hello {name}"},
@@ -214,12 +205,12 @@ def test_caching_no_reread(tmp_path: Path) -> None:
 
 def test_reload_forces_rescan(tmp_path: Path) -> None:
     """AC #4: reload() forces re-scan from disk."""
-    _write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
+    write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {name}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     assert len(repo.list()) == 1
 
     # Modify file on disk
-    _write_yaml(
+    write_yaml(
         tmp_path / "t.yaml",
         [
             {"id": "t1", "template": "Hello {name}"},
@@ -250,7 +241,7 @@ def test_create_persists_entry(tmp_path: Path) -> None:
 def test_create_appends_to_existing_file(tmp_path: Path) -> None:
     """create() appends to existing {id}.yaml file if it already exists."""
     target = tmp_path / "shared.yaml"
-    _write_yaml(target, [{"id": "shared", "template": "First {x}"}])
+    write_yaml(target, [{"id": "shared", "template": "First {x}"}])
     # create() writes to {id}.yaml — create an entry whose id matches the file
     repo = YamlTemplateCatalogRepository(tmp_path)
     entry = TemplateEntry(id="new-entry", template="Second {y}")
@@ -264,7 +255,7 @@ def test_create_appends_to_existing_file(tmp_path: Path) -> None:
 
 def test_create_duplicate_id_raises(tmp_path: Path) -> None:
     """create() raises CatalogValidationError if id already exists."""
-    _write_yaml(tmp_path / "existing.yaml", [{"id": "dup", "template": "T {x}"}])
+    write_yaml(tmp_path / "existing.yaml", [{"id": "dup", "template": "T {x}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     with pytest.raises(CatalogValidationError) as exc_info:
         repo.create(TemplateEntry(id="dup", template="New {y}"))
@@ -276,7 +267,7 @@ def test_create_duplicate_id_raises(tmp_path: Path) -> None:
 
 def test_update_modifies_entry(tmp_path: Path) -> None:
     """AC #3: update() modifies entry in file."""
-    _write_yaml(tmp_path / "t1.yaml", [{"id": "t1", "template": "Old {x}"}])
+    write_yaml(tmp_path / "t1.yaml", [{"id": "t1", "template": "Old {x}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     updated = TemplateEntry(id="t1", template="Updated {y}")
     repo.update("t1", updated)
@@ -288,7 +279,7 @@ def test_update_modifies_entry(tmp_path: Path) -> None:
 
 def test_update_single_dict_yaml(tmp_path: Path) -> None:
     """update() handles YAML file containing a single dict (not a list)."""
-    _write_yaml(tmp_path / "solo.yaml", {"id": "solo", "template": "Old {x}"})
+    write_yaml(tmp_path / "solo.yaml", {"id": "solo", "template": "Old {x}"})
     repo = YamlTemplateCatalogRepository(tmp_path)
     repo.update("solo", TemplateEntry(id="solo", template="New {y}"))
     entry = repo.get("solo")
@@ -298,7 +289,7 @@ def test_update_single_dict_yaml(tmp_path: Path) -> None:
 
 def test_update_raises_for_missing_id(tmp_path: Path) -> None:
     """update() raises EntryNotFoundError for missing id."""
-    _write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {x}"}])
+    write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {x}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     with pytest.raises(EntryNotFoundError):
         repo.update("nonexistent", TemplateEntry(id="nonexistent", template="T {x}"))
@@ -309,7 +300,7 @@ def test_update_raises_for_missing_id(tmp_path: Path) -> None:
 
 def test_delete_removes_entry_and_deletes_file_if_empty(tmp_path: Path) -> None:
     """AC #3: delete() removes entry, deletes file if empty."""
-    _write_yaml(tmp_path / "t1.yaml", [{"id": "t1", "template": "Hello {x}"}])
+    write_yaml(tmp_path / "t1.yaml", [{"id": "t1", "template": "Hello {x}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     repo.delete("t1")
     assert repo.get("t1") is None
@@ -318,7 +309,7 @@ def test_delete_removes_entry_and_deletes_file_if_empty(tmp_path: Path) -> None:
 
 def test_delete_keeps_file_with_remaining_entries(tmp_path: Path) -> None:
     """delete() removes entry but keeps file if other entries remain."""
-    _write_yaml(
+    write_yaml(
         tmp_path / "multi.yaml",
         [
             {"id": "a", "template": "A {x}"},
@@ -334,7 +325,7 @@ def test_delete_keeps_file_with_remaining_entries(tmp_path: Path) -> None:
 
 def test_delete_single_dict_yaml(tmp_path: Path) -> None:
     """delete() handles YAML file containing a single dict (not a list)."""
-    _write_yaml(tmp_path / "solo.yaml", {"id": "solo", "template": "T {x}"})
+    write_yaml(tmp_path / "solo.yaml", {"id": "solo", "template": "T {x}"})
     repo = YamlTemplateCatalogRepository(tmp_path)
     repo.delete("solo")
     assert repo.get("solo") is None
@@ -343,7 +334,7 @@ def test_delete_single_dict_yaml(tmp_path: Path) -> None:
 
 def test_delete_raises_for_missing_id(tmp_path: Path) -> None:
     """delete() raises EntryNotFoundError for missing id."""
-    _write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {x}"}])
+    write_yaml(tmp_path / "t.yaml", [{"id": "t1", "template": "Hello {x}"}])
     repo = YamlTemplateCatalogRepository(tmp_path)
     with pytest.raises(EntryNotFoundError):
         repo.delete("nonexistent")
