+++ tests/services/test_team_catalog.py
@@ -1,233 +1,18 @@
 """Tests for TeamCatalog service with cross-validation and downstream wiring."""
 
-import builtins
-
 import pytest
 
-from akgentic.catalog.models.agent import AgentEntry
 from akgentic.catalog.models.errors import CatalogValidationError, EntryNotFoundError
-from akgentic.catalog.models.queries import (
-    AgentQuery,
-    TeamQuery,
-    TemplateQuery,
-    ToolQuery,
-)
-from akgentic.catalog.models.team import TeamMemberSpec, TeamSpec
-from akgentic.catalog.models.template import TemplateEntry
-from akgentic.catalog.models.tool import ToolEntry
-from akgentic.catalog.repositories.base import (
-    AgentCatalogRepository,
-    TeamCatalogRepository,
-    TemplateCatalogRepository,
-    ToolCatalogRepository,
-)
+from akgentic.catalog.models.queries import TeamQuery
+from akgentic.catalog.models.team import TeamMemberSpec
 from akgentic.catalog.services.agent_catalog import AgentCatalog
 from akgentic.catalog.services.team_catalog import TeamCatalog
 from akgentic.catalog.services.template_catalog import TemplateCatalog
 from akgentic.catalog.services.tool_catalog import ToolCatalog
-
-_list = builtins.list
-
-
-# --- In-memory repositories for testing ---
-
-
-class InMemoryTemplateCatalogRepository(TemplateCatalogRepository):
-    def __init__(self) -> None:
-        self._entries: dict[str, TemplateEntry] = {}
-
-    def create(self, template_entry: TemplateEntry) -> str:
-        self._entries[template_entry.id] = template_entry
-        return template_entry.id
-
-    def get(self, id: str) -> TemplateEntry | None:
-        return self._entries.get(id)
-
-    def list(self) -> _list[TemplateEntry]:
-        return _list(self._entries.values())
-
-    def search(self, query: TemplateQuery) -> _list[TemplateEntry]:
-        return self.list()
-
-    def update(self, id: str, template_entry: TemplateEntry) -> None:
-        self._entries[id] = template_entry
-
-    def delete(self, id: str) -> None:
-        del self._entries[id]
-
-
-class InMemoryToolCatalogRepository(ToolCatalogRepository):
-    def __init__(self) -> None:
-        self._entries: dict[str, ToolEntry] = {}
-
-    def create(self, tool_entry: ToolEntry) -> str:
-        self._entries[tool_entry.id] = tool_entry
-        return tool_entry.id
-
-    def get(self, id: str) -> ToolEntry | None:
-        return self._entries.get(id)
-
-    def list(self) -> _list[ToolEntry]:
-        return _list(self._entries.values())
-
-    def search(self, query: ToolQuery) -> _list[ToolEntry]:
-        return self.list()
-
-    def update(self, id: str, tool_entry: ToolEntry) -> None:
-        self._entries[id] = tool_entry
-
-    def delete(self, id: str) -> None:
-        del self._entries[id]
-
-
-class InMemoryAgentCatalogRepository(AgentCatalogRepository):
-    def __init__(self) -> None:
-        self._entries: dict[str, AgentEntry] = {}
-
-    def create(self, agent_entry: AgentEntry) -> str:
-        self._entries[agent_entry.id] = agent_entry
-        return agent_entry.id
-
-    def get(self, id: str) -> AgentEntry | None:
-        return self._entries.get(id)
-
-    def list(self) -> _list[AgentEntry]:
-        return _list(self._entries.values())
-
-    def search(self, query: AgentQuery) -> _list[AgentEntry]:
-        return self.list()
-
-    def update(self, id: str, agent_entry: AgentEntry) -> None:
-        self._entries[id] = agent_entry
-
-    def delete(self, id: str) -> None:
-        del self._entries[id]
-
-
-class InMemoryTeamCatalogRepository(TeamCatalogRepository):
-    def __init__(self) -> None:
-        self._entries: dict[str, TeamSpec] = {}
-
-    def create(self, team_spec: TeamSpec) -> str:
-        self._entries[team_spec.id] = team_spec
-        return team_spec.id
-
-    def get(self, id: str) -> TeamSpec | None:
-        return self._entries.get(id)
-
-    def list(self) -> _list[TeamSpec]:
-        return _list(self._entries.values())
-
-    def search(self, query: TeamQuery) -> _list[TeamSpec]:
-        return self.list()
-
-    def update(self, id: str, team_spec: TeamSpec) -> None:
-        self._entries[id] = team_spec
-
-    def delete(self, id: str) -> None:
-        del self._entries[id]
-
-
-# --- Helper factories ---
-
-
-def _make_template(
-    id: str = "sys-prompt",
-    template: str = "You are {role}. {instructions}",
-) -> TemplateEntry:
-    return TemplateEntry(id=id, template=template)
-
-
-def _make_tool(id: str = "search-1") -> ToolEntry:
-    return ToolEntry(
-        id=id,
-        tool_class="akgentic.tool.search.search.SearchTool",
-        tool={"name": "search", "description": "Search the web"},
-    )
-
-
-def _make_agent(
-    id: str = "agent-1",
-    name: str = "test-agent",
-    tool_ids: _list[str] | None = None,
-    template_ref: str | None = None,
-    params: dict[str, str] | None = None,
-    routes_to: _list[str] | None = None,
-) -> AgentEntry:
-    prompt: dict[str, str | dict[str, str]] = {}
-    if template_ref is not None:
-        prompt["template"] = template_ref
-        if params is not None:
-            prompt["params"] = params
-    config: dict[str, str | dict[str, str | dict[str, str]]] = {"name": name}
-    if prompt:
-        config["prompt"] = prompt
-    return AgentEntry(
-        id=id,
-        tool_ids=tool_ids or [],
-        card={
-            "role": "engineer",
-            "description": "test agent",
-            "skills": ["coding"],
-            "agent_class": "akgentic.agent.BaseAgent",
-            "config": config,
-            "routes_to": routes_to or [],
-        },
-    )
-
-
-def _make_team(
-    id: str = "team-1",
-    name: str = "Test Team",
-    entry_point: str = "agent-1",
-    members: _list[TeamMemberSpec] | None = None,
-    profiles: _list[str] | None = None,
-    message_types: _list[str] | None = None,
-) -> TeamSpec:
-    default_members = members or [TeamMemberSpec(agent_id="agent-1")]
-    return TeamSpec(
-        id=id,
-        name=name,
-        entry_point=entry_point,
-        message_types=message_types or ["akgentic.core.messages.UserMessage"],
-        members=default_members,
-        profiles=profiles or [],
-    )
-
+from tests.conftest import make_agent, make_team, make_template, make_tool
+from tests.services.conftest import InMemoryAgentCatalogRepository, InMemoryTeamCatalogRepository
 
 # --- Fixtures ---
-
-
-@pytest.fixture
-def template_repo() -> InMemoryTemplateCatalogRepository:
-    return InMemoryTemplateCatalogRepository()
-
-
-@pytest.fixture
-def tool_repo() -> InMemoryToolCatalogRepository:
-    return InMemoryToolCatalogRepository()
-
-
-@pytest.fixture
-def agent_repo() -> InMemoryAgentCatalogRepository:
-    return InMemoryAgentCatalogRepository()
-
-
-@pytest.fixture
-def team_repo() -> InMemoryTeamCatalogRepository:
-    return InMemoryTeamCatalogRepository()
-
-
-@pytest.fixture
-def template_catalog(
-    template_repo: InMemoryTemplateCatalogRepository,
-) -> TemplateCatalog:
-    return TemplateCatalog(repository=template_repo)
-
-
-@pytest.fixture
-def tool_catalog(tool_repo: InMemoryToolCatalogRepository) -> ToolCatalog:
-    return ToolCatalog(repository=tool_repo)
 
 
 @pytest.fixture
@@ -242,9 +27,9 @@ def agent_catalog(
         tool_catalog=tool_catalog,
     )
     # Pre-populate with known agents for team validation
-    cat.create(_make_agent("agent-1", name="agent-one"))
-    cat.create(_make_agent("agent-2", name="agent-two"))
-    cat.create(_make_agent("agent-3", name="agent-three"))
+    cat.create(make_agent("agent-1", name="agent-one"))
+    cat.create(make_agent("agent-2", name="agent-two"))
+    cat.create(make_agent("agent-3", name="agent-three"))
     return cat
 
 
@@ -263,14 +48,14 @@ class TestValidateCreateDuplicateId:
     """AC5: Duplicate ID rejection."""
 
     def test_returns_error_for_duplicate_id(self, catalog: TeamCatalog) -> None:
-        team = _make_team("team-1")
+        team = make_team("team-1")
         catalog.create(team)
-        errors = catalog.validate_create(_make_team("team-1"))
+        errors = catalog.validate_create(make_team("team-1"))
         assert len(errors) == 1
         assert "already exists" in errors[0]
 
     def test_returns_empty_for_unique_id(self, catalog: TeamCatalog) -> None:
-        errors = catalog.validate_create(_make_team("team-1"))
+        errors = catalog.validate_create(make_team("team-1"))
         assert errors == []
 
 
@@ -278,12 +63,12 @@ class TestValidateCreateValidTeam:
     """AC1: Valid team creation with all agents existing."""
 
     def test_valid_team_no_errors(self, catalog: TeamCatalog) -> None:
-        team = _make_team("team-1", members=[TeamMemberSpec(agent_id="agent-1")])
+        team = make_team("team-1", members=[TeamMemberSpec(agent_id="agent-1")])
         errors = catalog.validate_create(team)
         assert errors == []
 
     def test_valid_team_multiple_members(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             members=[TeamMemberSpec(agent_id="agent-1"), TeamMemberSpec(agent_id="agent-2")],
         )
@@ -295,7 +80,7 @@ class TestValidateCreateEntryPoint:
     """AC2: Entry point validation."""
 
     def test_entry_point_not_in_members(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="agent-999",
             members=[TeamMemberSpec(agent_id="agent-1")],
@@ -304,7 +89,7 @@ class TestValidateCreateEntryPoint:
         assert any("Entry point 'agent-999' not found in members tree" in e for e in errors)
 
     def test_entry_point_found_in_nested_members(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="agent-2",
             members=[TeamMemberSpec(
@@ -316,7 +101,7 @@ class TestValidateCreateEntryPoint:
         assert not any("Entry point" in e for e in errors)
 
     def test_entry_point_agent_not_in_catalog(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="nonexistent",
             members=[TeamMemberSpec(agent_id="nonexistent")],
@@ -329,7 +114,7 @@ class TestValidateCreateMemberAgents:
     """AC15: Member agent_id doesn't exist in AgentCatalog."""
 
     def test_missing_member_agent(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="agent-1",
             members=[TeamMemberSpec(agent_id="agent-1"), TeamMemberSpec(agent_id="nonexistent")],
@@ -338,7 +123,7 @@ class TestValidateCreateMemberAgents:
         assert any("Agent 'nonexistent' not found in AgentCatalog" in e for e in errors)
 
     def test_nested_missing_member_agent(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="agent-1",
             members=[TeamMemberSpec(
@@ -354,7 +139,7 @@ class TestValidateCreateProfiles:
     """AC3: Profile agent validation."""
 
     def test_missing_profile_agent(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             members=[TeamMemberSpec(agent_id="agent-1")],
             profiles=["nonexistent"],
@@ -363,7 +148,7 @@ class TestValidateCreateProfiles:
         assert any("Profile agent 'nonexistent' not found in AgentCatalog" in e for e in errors)
 
     def test_valid_profile_agent(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             members=[TeamMemberSpec(agent_id="agent-1")],
             profiles=["agent-2"],
@@ -376,7 +161,7 @@ class TestValidateCreateMessageTypes:
     """AC4: Message type resolution."""
 
     def test_unresolvable_message_type(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             message_types=["nonexistent.module.FakeClass"],
         )
@@ -384,7 +169,7 @@ class TestValidateCreateMessageTypes:
         assert any("Cannot resolve message_type" in e for e in errors)
 
     def test_valid_message_type(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             message_types=["akgentic.core.messages.UserMessage"],
         )
@@ -396,7 +181,7 @@ class TestValidateCreateMultipleErrors:
     """AC6: Multiple error collection in single call."""
 
     def test_multiple_errors_collected(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="missing-ep",
             members=[TeamMemberSpec(agent_id="nonexistent-member")],
@@ -420,13 +205,13 @@ class TestCreate:
     """AC6: create delegates to validate_create and raises."""
 
     def test_create_success(self, catalog: TeamCatalog) -> None:
-        team = _make_team("team-1")
+        team = make_team("team-1")
         result = catalog.create(team)
         assert result == "team-1"
         assert catalog.get("team-1") is not None
 
     def test_create_raises_on_validation_errors(self, catalog: TeamCatalog) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="nonexistent",
             members=[TeamMemberSpec(agent_id="nonexistent")],
@@ -442,7 +227,7 @@ class TestCrudDelegation:
     """AC8: get, list, search delegate directly to repository."""
 
     def test_get_existing(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
+        catalog.create(make_team("team-1"))
         result = catalog.get("team-1")
         assert result is not None
         assert result.id == "team-1"
@@ -451,8 +236,8 @@ class TestCrudDelegation:
         assert catalog.get("nonexistent") is None
 
     def test_list_delegates(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
-        team2 = _make_team(
+        catalog.create(make_team("team-1"))
+        team2 = make_team(
             "team-2", entry_point="agent-2", members=[TeamMemberSpec(agent_id="agent-2")],
         )
         catalog.create(team2)
@@ -460,7 +245,7 @@ class TestCrudDelegation:
         assert len(result) == 2
 
     def test_search_delegates(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
+        catalog.create(make_team("team-1"))
         query = TeamQuery(id="team-1")
         result = catalog.search(query)
         assert len(result) >= 1
@@ -473,29 +258,29 @@ class TestUpdate:
     """AC7: update with existence check and cross-validation."""
 
     def test_update_existing_success(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
-        updated = _make_team("team-1", name="Updated Team")
+        catalog.create(make_team("team-1"))
+        updated = make_team("team-1", name="Updated Team")
         catalog.update("team-1", updated)
         result = catalog.get("team-1")
         assert result is not None
         assert result.name == "Updated Team"
 
     def test_update_nonexistent_raises(self, catalog: TeamCatalog) -> None:
-        team = _make_team("team-1")
+        team = make_team("team-1")
         with pytest.raises(EntryNotFoundError, match="not found"):
             catalog.update("team-1", team)
 
     def test_update_id_mismatch_raises(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
-        mismatched = _make_team(
+        catalog.create(make_team("team-1"))
+        mismatched = make_team(
             "team-2", entry_point="agent-2", members=[TeamMemberSpec(agent_id="agent-2")],
         )
         with pytest.raises(CatalogValidationError, match="does not match"):
             catalog.update("team-1", mismatched)
 
     def test_update_cross_validates(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
-        updated = _make_team(
+        catalog.create(make_team("team-1"))
+        updated = make_team(
             "team-1",
             members=[TeamMemberSpec(agent_id="nonexistent")],
         )
@@ -503,8 +288,8 @@ class TestUpdate:
             catalog.update("team-1", updated)
 
     def test_update_skips_own_duplicate_check(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
-        updated = _make_team("team-1", name="Renamed Team")
+        catalog.create(make_team("team-1"))
+        updated = make_team("team-1", name="Renamed Team")
         catalog.update("team-1", updated)
         result = catalog.get("team-1")
         assert result is not None
@@ -518,7 +303,7 @@ class TestDelete:
     """AC14: Delete with no downstream refs (v1: teams are independent)."""
 
     def test_delete_success(self, catalog: TeamCatalog) -> None:
-        catalog.create(_make_team("team-1"))
+        catalog.create(make_team("team-1"))
         catalog.delete("team-1")
         assert catalog.get("team-1") is None
 
@@ -538,9 +323,9 @@ class TestDownstreamWiringTemplateDelete:
         template_catalog: TemplateCatalog,
         agent_catalog: AgentCatalog,
     ) -> None:
-        template_catalog.create(_make_template("sys-prompt"))
+        template_catalog.create(make_template("sys-prompt"))
         # Create agent that references template
-        agent = _make_agent(
+        agent = make_agent(
             "ref-agent",
             name="ref-agent",
             template_ref="@sys-prompt",
@@ -556,7 +341,7 @@ class TestDownstreamWiringTemplateDelete:
         self,
         template_catalog: TemplateCatalog,
     ) -> None:
-        template_catalog.create(_make_template("sys-prompt"))
+        template_catalog.create(make_template("sys-prompt"))
         assert template_catalog.agent_catalog is None
         template_catalog.delete("sys-prompt")
         assert template_catalog.get("sys-prompt") is None
@@ -570,8 +355,8 @@ class TestDownstreamWiringToolDelete:
         tool_catalog: ToolCatalog,
         agent_catalog: AgentCatalog,
     ) -> None:
-        tool_catalog.create(_make_tool("search-1"))
-        agent = _make_agent("ref-agent", name="ref-agent", tool_ids=["search-1"])
+        tool_catalog.create(make_tool("search-1"))
+        agent = make_agent("ref-agent", name="ref-agent", tool_ids=["search-1"])
         agent_catalog.create(agent)
         # Wire downstream
         tool_catalog.agent_catalog = agent_catalog
@@ -582,7 +367,7 @@ class TestDownstreamWiringToolDelete:
         self,
         tool_catalog: ToolCatalog,
     ) -> None:
-        tool_catalog.create(_make_tool("search-1"))
+        tool_catalog.create(make_tool("search-1"))
         assert tool_catalog.agent_catalog is None
         tool_catalog.delete("search-1")
         assert tool_catalog.get("search-1") is None
@@ -596,7 +381,7 @@ class TestDownstreamWiringAgentDeleteByTeamMember:
         agent_catalog: AgentCatalog,
         catalog: TeamCatalog,
     ) -> None:
-        team = _make_team("team-1", members=[TeamMemberSpec(agent_id="agent-1")])
+        team = make_team("team-1", members=[TeamMemberSpec(agent_id="agent-1")])
         catalog.create(team)
         agent_catalog.team_catalog = catalog
         with pytest.raises(CatalogValidationError, match="cannot delete"):
@@ -611,7 +396,7 @@ class TestDownstreamWiringAgentDeleteByTeamProfile:
         agent_catalog: AgentCatalog,
         catalog: TeamCatalog,
     ) -> None:
-        team = _make_team(
+        team = make_team(
             "team-1",
             members=[TeamMemberSpec(agent_id="agent-1")],
             profiles=["agent-2"],
@@ -631,7 +416,7 @@ class TestDownstreamWiringAgentDeleteByRouting:
     ) -> None:
         # agent-1 already exists with name "agent-one"
         # Create agent that routes to "agent-one"
-        router = _make_agent("router", name="router-name", routes_to=["agent-one"])
+        router = make_agent("router", name="router-name", routes_to=["agent-one"])
         agent_catalog.create(router)
         with pytest.raises(CatalogValidationError, match="cannot delete"):
             agent_catalog.delete("agent-1")
@@ -652,7 +437,7 @@ class TestDownstreamWiringStandaloneUsage:
         self,
         template_catalog: TemplateCatalog,
     ) -> None:
-        template_catalog.create(_make_template("sys-prompt"))
+        template_catalog.create(make_template("sys-prompt"))
         assert template_catalog.agent_catalog is None
         template_catalog.delete("sys-prompt")
         assert template_catalog.get("sys-prompt") is None
@@ -661,7 +446,7 @@ class TestDownstreamWiringStandaloneUsage:
         self,
         tool_catalog: ToolCatalog,
     ) -> None:
-        tool_catalog.create(_make_tool("search-1"))
+        tool_catalog.create(make_tool("search-1"))
         assert tool_catalog.agent_catalog is None
         tool_catalog.delete("search-1")
         assert tool_catalog.get("search-1") is None
@@ -684,11 +469,11 @@ class TestFullFourCatalogWiring:
         agent_catalog.team_catalog = catalog
 
         # 3. Create a template and tool
-        template_catalog.create(_make_template("sys-prompt"))
-        tool_catalog.create(_make_tool("search-1"))
+        template_catalog.create(make_template("sys-prompt"))
+        tool_catalog.create(make_tool("search-1"))
 
         # 4. Create an agent referencing both
-        ref_agent = _make_agent(
+        ref_agent = make_agent(
             "ref-agent",
             name="ref-agent",
             tool_ids=["search-1"],
@@ -698,7 +483,7 @@ class TestFullFourCatalogWiring:
         agent_catalog.create(ref_agent)
 
         # 5. Create a team referencing the agent
-        team = _make_team(
+        team = make_team(
             "team-1",
             entry_point="ref-agent",
             members=[TeamMemberSpec(agent_id="ref-agent")],
